+++ frontend/src/AppRoutes.tsx
@@ -2,7 +2,7 @@ import { Route, Routes } from 'react-router-dom';
 import { Box, styled } from '@mui/system';
 import React, { ReactNode, useEffect } from 'react';
 import { useNavigate } from 'react-router';
-import { Home, Login, SignUp } from './views';
+import { Home, Login, SignUp } from './pages';
 import { logout } from './utils';
 import AppBar from '@mui/material/AppBar';
