+++ frontend/src/__tests__/home.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import { Home } from '../views/Home';
+import { Home } from '../pages/Home';
 
 it('Home renders correctly', () => {
   const home = render(<Home />);
