+++ frontend/src/__tests__/login.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import { Login } from '../views';
+import { Login } from '../pages';
 
 it('Login renders correctly', () => {
   const login = render(<Login />);
