+++ frontend/src/index.css
@@ -257,3 +257,13 @@ body {
 .home-page-body {
     margin: 30px 120px;
 }
+@media screen and (max-width: 750px) {
+  .homepage {
+    gap: 50px 100px;
+  }
+}
+@media screen and (max-width: 450px) {
+  .homepage {
+    gap: 25px 100px;
+  }
+}
